Создал файл Application, добавл несколько новых классов

+++ src/Clients/Cat.java
@@ -2,7 +2,7 @@ package Clients;
 
 
 public class Cat extends Animal{
-    public Cat(String name, float weight, java.time.LocalDate age, Owner owner){
+    public Cat(String name, double weight, java.time.LocalDate age, Owner owner){
         super(name, weight, age, owner);
 
     }
