+++ src/Clients/Dog.java
@@ -2,7 +2,7 @@ package Clients;
 
 
 public class Dog extends Animal{
-    public Dog(String name, float weight, java.time.LocalDate age, Owner owner){
+    public Dog(String name, double weight, java.time.LocalDate age, Owner owner){
         super(name, weight, age, owner);
 
     }
