+++ src/Clients/Lion.java
@@ -1,12 +1,12 @@
 package Clients;
 
 public class Lion extends Animal{
-    public Lion (String name, float weight, java.time.LocalDate age, Owner owner){
+
+    public Lion (String name, double weight, java.time.LocalDate age, Owner owner){
         super(name, weight, age, owner);
     }
     
     public Lion(){
         super();
     }
-
 }
